Update README.md
Bitcoin-investors
Invest with us 
I have software that generate over $5000 in 1 week on binary trading option with amazing strategies and skills i employ i manage and guide you on how to earn up to $5000 successful 
1 week. feel free to contact me if you are interested to earn extra profits on investing 
ATTENTION EVERYBODY
Are you interested in binary trading,invest with a trusted account manager and get a better return in 7 days,i can help you manage your account with the minimum of the following without wasting time
$200 to earn $2,000
$300 to earn $3,000
$500 to earn $5,000
$600 to earn $6,000
$1,000 to earn $10,000
$2,000 to earn $20,000
$3,000 to earn $30,000
$4,000 to earn $40,000
$5,000 to earn $50,000
$6,000 to earn $60,000
$7,000 to earn $70,000
$8,000 to earn $80,000
$9,000 to earn $90,000
Interested people should inbox me "INFO
After posting the above when they message
Chat introduction
Hello
How are you doing?
Where are you originally from?
Are you interested in earning bitcoin
Have you heard about bitcoin and binary option trade?
Bitcoin is a kind of money that can be sent from one person to another without the need for a trusted third party such as a bank or financial institution,it is the first global,decentralized currency.
Investment works more like a prediction that can be either right or wrong,but no where in between.An investor bets that a given asset will be above or below a certain point after a fixed ammount of time. 
So to avoid the loose of your investment you have to trade with an expert of an account manager in trading like me.
You can trust the secure connection,reliability and speed of our multiple award-winning platforms technology,Furthermore our platform is integrated with a range of advanced tools meant to guide traders in charting,trading and technical analysis.